[WIP] Add Windows SMB share driver
This patch proposes a new share driver, handling Windows service
instances and exporting SMB shares.

WinRM is used for managing the service instances, allowing
password or certificate based authentication.

A new service instance manager is introduced, handling Windows
instances. One requirement is that Cloudbase-Init is present on the
instance.

If AD credentials are provided via security services, the instance
is joined to an AD domain. This can be leveraged in order to set
user based permissions.

Unit tests will be included soon. Also, we are working on providing
a CI for this driver. In the absence of it, Tempest results will be
provided. Note that this work will be splitted in multiple patch
sets.

Meanwhile, this is marked as WIP.

Change-Id: Ice5ee63ee4dc2cfcb41567214c165c2705421920
Implements: blueprint windows-smb-support