Windows SMB: avoid powershell errors when fetching share locations
When fetching SMB shares that do not exist, we expect a PowerShell
error to be thrown, which we suppress if we expect it to be missing.

This will pollute the logs, as the expected error is being logged.
This patch avoids this situation by avoiding a powershell error to
be thrown.

Change-Id: I7776cca806a59f55e80cbd38d0fb7bb09861228a